feat(roles): update role parsing to use parseFloat for better handling of decimal roles feat(unisym): add extractCommandRole function to streamline role extraction logic fix(middleware): refactor command role extraction to utilize extractCommandRole fix(global): add databases property to global Cassidy object for better database management fix(package): bump version to 3.5.9-PRERELEASE

+++ CommandFiles/commands/roles.ts
@@ -110,7 +110,7 @@ const configs: Config[] = [
       { spectralArgs }
     ) {
       const commandName = spectralArgs[0];
-      const role = parseInt(spectralArgs[1]);
+      const role = parseFloat(spectralArgs[1]);
       const isGlobal = spectralArgs[2] === "-g";
 
       const targetCommand = Object.values(commands).find(
